v4l2/stateless: Configure output format for pipeline
The format of the |output_queue_| has already been negotiated
with the driver. This output format is then passed to
consumers further down the pipeline.

This chosen format may not be useable as is. Check with
and configure the pipeline to be able to consume the format
produced by the driver.

Bug: b:278935312

Change-Id: I123dadfc2c4bfbc01a67230c0dee475e96d01036
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5069635
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231020}