Move form_data fuzz productor to /autofill/
Refactor the password_manager's fuzzer helpers to decouple the pure
autofill FormData generation, and move the latter to
//components/autofill/.

Follow-ups will use this in order to enhance autofill's fuzzers.

Note: The side effect of this refactoring is that the data generation
consumes more of the input bytes, because some low-level bitwise tricks
(reading booleans from specific bits of input bytes) ended up in
different functions and hence each of them needs a byte on its own.
However, we assume it's not a huge deal, since it's only extra 15 bytes
in total.

Bug: 1505445
Change-Id: I8fb78bcc51e8d9960707cf4dbf451fc1b4d05a63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5063015
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Vaclav Brozek <vabr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231024}