Set PDF to SAMPLED_PING
This file type is a huge amount of traffic but a very small amount of
warnings. It's not a good tradeoff to be sending download pings for
this file type. By setting it's file type policy to SAMPLED_PING, we
will only contact Safe Browsing about a small fraction of downloads.
This preserves some visibility (in case we need to revert this),
without paying the ongoing cost of delaying user downloads.

Bug: 1495557
Change-Id: Ic5eb7fca22078be984b3b34f48c731adad88ab05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4960970
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Christine Smith <christinesm@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231035}