Remove unused callback_ field from ash::UnifiedSliderView.
This was not used before, but it was harmless, since PressedCallback is
currently copyable. However, PressedCallback will be changed to be
compatible with OnceCallback, at which point it will no longer be
copyable. Since this copy isn't needed, removing it is the simplest fix
here.

Bug: 1326449
Change-Id: I5c8d8c88cb0cfd8f2db3721d8d99b9ad7da7d44c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5074395
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231047}