Abstract OpenXR Hand Joint Data parsing
Move the OpenXR HandTracker logic out of openxr_controller and into a
delegated class. This helps to encapsulate the various static asserts
that were being made there, as well as provides an avenue towards
potential future flexibility to parse gestures out from additional
hand tracker related extensions. (https://crrev.com/c/5047184 is one
such example of a proposal to do this.)

Also updates and clarifies the requirements for enabling the hand-input
feature; namely that we need the XR_EXT_HAND_TRACKING extension to be
enabled to get the hand data *and* we need an extension that can
generate a "primary interaction" for that hand to be enabled (of which
there is currently only the XR_MSFT_HAND_INTERACTION).

Bug: 1504076
Change-Id: Iecfb8b6a3e2a84ef87f5b239b58caf96799fa413
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5072895
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231074}