AW: rename unittest
No change to logic. This renames a unittest. The package name allowlist
moved to server side so it doesn't make sense for this unittest to have
"allowlist" in the name. This keeps the test since it asserts that
AwMetricsServiceClient returns `true` as it always should.

Test: N/A
Change-Id: Ic12ff42b1290f78f2724176f4bacf694f801e9ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5074453
Reviewed-by: Adam Walls <avvall@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231089}