Make the data pipe size in RaceNetworkRequest configurable
In ServiceWorkerRaceNetworkRequestURLLoaderClient, two mojo pipes are
newly created in the constructor. If the ServiceWorkerAutoPreload
feature is enabled, this happens basically per resource requests,
inclusding subresources.

We're seeing SIGBUS error increase on Linux dev channel, in which we run
the ServiceWorkerAutoPreload experiment now. This may happen on the
enviroment having small /dev/shm space.

This CL makes the buffer size of the data pipe configurable so that we
can check if the resource usage of mojo data pipe causes the error or
not, by passing the smaller number as a buffer size for the data pipe
creation via the feature param.

Change-Id: I455cf9693bd74f159f1e4afb46116b76c6cc059e
Bug: 1502946
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5069536
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Auto-Submit: Shunya Shishido <sisidovski@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231151}