Add display color management types
This is towards combining all color management data closer to
DrmDevice (in the viz process).

The type display::ColorCalibration will replace
ash::DisplayColorManager::ColorCalibrationData. It may be that the internals of this structure will be replaced with an SkColorSpace
eventually.

The type display::ColorTemperatureAdjustment will be used for the
night light matrix. It is getting its own structure (as opposed
to just using a skcms_Matrix3x3) because the implementation may be
simplified to reflect that it (currently) is always a diagonal matrix.

The type display::GammaAdjustment will be used to communicate the
gamma changes from chromecast/. Those currently follow the same
flow as the color calibration despite having very different
semantics.

Add these types, their mojo equivalents, and tests.

Bug: 1505062
Change-Id: Id19c7891193c65043797741a0cd648f710910110
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5062933
Reviewed-by: Rick Byers <rbyers@chromium.org>
Reviewed-by: Brian Osman <brianosman@google.com>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231160}