[ChromeSigninBubble] Fix round corners
The bubble corners were actually round, however the loaded web view
was overflowing on the edges (rounded corners).
This was due to the fact that were overriding the default margins of
the bubbles.

To fix this, we keep the default margins, and adapt the webui
padding to get the desired space. The values of the padding
are different with/without Chrome Refresh 2023 because the
default value of the bubble margins are different with/without
Chrome Refresh 2023.

Before fix screenshot: https://screenshot.googleplex.com/4t78Upcs3phhhfQ
(where the overlap can be seen with the dark background)
After fix screenshot: https://screenshot.googleplex.com/6USFUzHd37NBJdr

Bug: b:306585843
Change-Id: I3e802bc095399ccb59c5f0a498bcd692474d12f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5057551
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Gabriel Oliveira <gabolvr@google.com>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1231238}