[mullet m3] Don't test BAM in the AED tests.
The unit tests for the AutofillExternalDelegate set expectations on
the BAM -> AutofillDriver interactions. This CL removes this by
mocking BAM::FillOrPreviewField and setting expectations on
AED -> BAM interactions.

Mocked function AutofillDriver::ApplyFieldAction is removed as not
used as a result of this CL.

Bug: 1459990
Change-Id: I3d02abf1ccacec0df904d89471ad98ec7a19bef5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5077012
Reviewed-by: Jan Keitel <jkeitel@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Cr-Commit-Position: refs/heads/main@{#1231251}