Make in_session_auth.mojom interface implementation shareable
We move the in_session_auth_ash implementation of the interface to
chromeos/components/in_session_auth/

The aim is to have this implementation in shareable code, so that
we can expose it to both crosapi, and chromeos settings webuis.

The instance lives in-process, and is a long lived object with static
storage, see
chromeos/components/in_session_auth/in_process_instances.cc. This
instance was previously owned by `crosapi_ash`.

Bug: b:259528315
Change-Id: I1258902493f6f035851a50a8edf746e0d8e755bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4987352
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Elie Maamari <emaamari@google.com>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231260}