Add workaround for last_active_time underlying type transition
Before https://crrev.com/c/4905251, last_active_time was stored
as a base::TimeTicks internal value. After the CL, it is stored
as microseconds since Windows epoch.
TimeTicks' internal representation (i.e. the old persisted value)
does *not* match "microseconds since Windows epoch", which
is leading to an issue where the values serialized with a
TimeTicks origin were deserialized very far in the past when
looked from a base::Time epoch.

This CL adds a code path that tries to detect if the value was stored
in the old format, and if yes transform it to the new format. Once
converted to the new format, the value will be re-serialized in the
new format.

It is a temporary workaround that should be removed once most users
have converted their last_active_time to the new format.

This is using banned functions as this is what was used previously
(see https://crrev.com/c/4905251).

Bug: 1490302, 1506068
Change-Id: I2e0c90aedfc5b479afa7fd801a05d83b998b7ccb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5071355
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231281}