Delete CrdHostSession when the session is terminated.
This makes it easier to reason about the state of the session
(it's there -> it is active, it's not there -> it is not active).

It also means that I can now add scoped objects to CrdHostSession
and they will behave as expected (in a follow-up CL).

Bug: b:310623115
Change-Id: I32829c08857dec0b8871b574152eda7d2e920a46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5054068
Reviewed-by: Ben Franz <bfranz@chromium.org>
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Cr-Commit-Position: refs/heads/main@{#1231335}