[pwm] Record filling assistance for single username forms.
This metric will give us an idea on how good is PWM at filling
the username form in the sign-in UFF flow. We will use that metric
to evaluate the filling assistance for single username forms
as we rollout sign-in UFF on iOS.

The metric is recorded for every single username field the user has
filled (automatically or manually), when the form is submitted and
the corresponding form manager is destroyed.

Change-Id: I1048a6073add2eaefdc525c24b0a0e5b4bdc4985
Bug: b/305313926
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4994684
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Maria Kazinova <kazinova@google.com>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#1231337}