[ios] Fix incorrect assertion when using multiple windows on iPad
On iPad, it is possible to have multiple windows, and thus Browsers,
open at the same time. This means it is possible for another Browser
to be dirty after loading the session for one Browser.

This means that it is possible for `dirty_web_state_lists_` to be
non-empty after removing the current Browser's WebStateList from it.
Handle this case by removing the DCHECK() and only cancelling the
timer if the list is empty.

Bug: 1383087
Change-Id: I0b54970f2f46e2e3a630b90eca541d869a200724
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5076379
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Federica Germinario <fedegermi@google.com>
Cr-Commit-Position: refs/heads/main@{#1231352}