[ChromeSigninBubble] Limit the number of times the bubble can be shown to the user
The information is stored in a pref that a map of email hashes to a
count: signin.ChromeSigninInterceptionShownCountPref. The maximum
number of times the bubble is allowed to be shown is hardcoded to 5,
and the count is incremented regardless of the result of the bubble.

Bug: b:313860458
Change-Id: I0a3b35fbcdf34dd82faa50d44192f97bc21e8ca6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5073972
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1231365}