Reliably record notification permission revocations.
First, record if there is a ScopedRevocationReporter on the stack in the onContentSettingsChanged path. If there is one, record which PermissionSourceUI it came from. If there isn't one, manually call PermissionRevoked and record an 'UNIDENTIFIED' PermissionSourceUI.

We suspect that on Android, UNIDENTIFIED actually maps to ANDROID_SETTINGS, since permission toggles in Android are not handled in Chrome directly (i.e. in the OS Settings), and differences are detected and observed via the onContentSettingsChanged method. In these cases there is no explicit ScopedRevocationReporter created which means our current logging misses these events.

Bug: 1473118
Change-Id: Ib6d0df909c22210d313a12db5d9d082b25376aee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4824826
Commit-Queue: William Xu <willxu@google.com>
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Reviewed-by: Ravjit Uppal <ravjit@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231368}