[ios] Use newer Chrome symbol for SetUpList default browser item
The newer symbol includes a white ring around the blue center. Because
the older symbol did not include this, a white background was added
behind the symbol so that it looked correct in dark mode - otherwise the
background would have turned dark. With the new symbol, we don't need
this complication.

Bug: 1489199
Change-Id: Ib4b038a30c95ee4145d31f4e4543dbff25396c9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5074622
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1231400}