Fix background colors of page info/action icons in PWA windows.
Make sure that the various kColorPageInfo* background colors are
correct for usage in the PWA title bar, rather than in an info-bar.
For the one case where such an icon is actually used in an info bar
(in CustomTabBarView), override GetLocationIconBackgroundColorOverride
to get the correct background color.

This does mean that the hover/pressed colors are slightly incorrect
in CustomTabBarView, but that is better than the regular state of icons
in the titlebar being completely wrong.

Bug: 1501817, b:312411708
Change-Id: If0a6e5625a99bf352c4eeafabb80e48530825b21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5068517
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231413}