[iOS] Fix Payments Bottom Sheet's GPay logo
The GPay logo was resized using a draw operation in order to make
its actual height 28 pixels (from 24). This introduced visual
artefacts because of the low quality of the resulting image.

Instead, we now adjust the image's ratio (which, for native images
is usually 3.0) to keep the pixel data (3 * 24 = 72 pixels high) as
is and instead draw that in a larger region on the screen. So, in
this case, adjusting the ratio from 3 to (24/28) * 3 = ~2.5714...
will display the 72 pixels high pixel data in a 28 pixels high
region on the screen. This will yield better results than actually
resizing the pixel data to be 28 pixels high (and will save a draw
operation).

Bug: 1503583
Change-Id: I9fbe0b6acd79f22493751a35e004b99d92cf3576
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5074732
Commit-Queue: Alexis Hétu <sugoi@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231426}