glanceables bar: add bg blur to show/hide buttons
This cl enables blurred background shield of show and hide chips
buttons on glanceables chips bar. After enabling it, the icon button
will add a blurred background shield which is a blurred solid color
layer stacked below the button's layer. However, it is very hard to
synchronize the resizing and opacity animations of the background
shield layer and button layer of show chips button. To solve this,
we put show chips button in a container view and animate the
container instead of the button.

Bug: b:313934294
Change-Id: I2322f3e864f9bdad3aa428e775f29cb083cb84dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5072885
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231432}