Remove mad@chromium.org from chrome/browser/extensions/api/enterprise_reporting_private/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/mad@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I2d33ae2ff755aee3b4eb2dc71d3fd158a7fa99ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5067221
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1231436}