Split magnifier effects into one backdrop zoom and one forward offset
The intent of backdrop filters was to be clipped to the layer's
bounds. When
https://chromium-review.googlesource.com/c/chromium/src/+/5002826
landed, it patched up a bug that was allowing certain cases to avoid
this clipping. Importantly, the touch selection magnifier was one of
these.

Assuming the backdrop filter could read outside of a layer's bounds,
the touch selection magnifier operated correctly. Discussing with
@kylechar, it seemed better to enforce that all backdrop filters keep
to this isolation rule than have a mode where UI layers can read
beyond their layer.

With that restriction, "backdrop offset" for the purposes of the
magnifier would be fundamentally broken because the content it needs
to magnify and where it has to be drawn are disjoint. To get around
this, the magnifier widget now places the "zoom" layer above the
region to be magnified by the Zoom background effect. Then it uses
a regular offset filter to translate everything to line up with the
rest of the magnifier widget's display, since forward filters are
allowed to move pixels from the layer bounds to elsewhere.

To get this to work, ui::Layer's background offset controls are
removed and replaced with a "layer offset" to reflect that the
cc::FilterOperation::OFFSET that it creates is now a regular filter
and not a backdrop filter.

ui::Layer now also exposes control over the cc::Layer's backdrop
filter bounds, which is a clip that impacts the backdrop filter
effect. Previously, SetRoundedCorners() was used, but this creates
a clip aligned with the original layer coordinates applied but applied
post-offset filter (so with a sufficient offset, it would clip out
everything of interest). Now the rounded corners are applied to the
output of the ZOOM backdrop effect before the offset filter and the
effect is carried along with it.

This also represents the first use case of a single RPDQ with
filters and backdrop filters, so there were several bugs and
simplifications that had to/could be made to SkiaRenderer.

Bug: b/312056981
Change-Id: I17967a659f9385415d8e186ae1d070c11cbd4da9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5046973
Reviewed-by: Michelle Chen <michellegc@google.com>
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Cr-Commit-Position: refs/heads/main@{#1231438}